Create and implement `ConditionForm`
The ConditionForm component will handle the user interaction in selecting the
specific condition fields of the statute.

Once the user has made their selections and submitted the form, these condition
field variations will be updated in the url query params.

The reason we do this instead of sending the data to the backend directly is
so that we can have persisted statute data even on refresh or manual url
navigation.