Add in useNavigate hook to be used in Body select
This hook was left out of the Select element in a previous commit.
In a normal project I would amend this to the original commit but since this is
a pet project and it would be a pain I am leaving it in its own commit.

If you are a future employer reading this know I am only doing this because its
a pet project and it's late and I am tired.