Add title column to  Statute and switch with name
The purpose of this is to have an internal name that we can use in the url
params instead of a generic number id. Now the statute name will be in the url.
This should improve the user experience.

As a result we need to run a our find queries using the name attribute instead
of the id attribute.

We also need to create a new column `title` for the displayed version of the
statute name.