Add check for if the Select input is valid
As a form of validation we want to disable the Update Sentence button if the
select input is invalid. Right now the only way it would be invalid is if the
input value is the same as the defaultValue which is 'false'.

Once a user makes a selection the input is now valid and the sentence can be
updated.