Remove strict typing dependent_condition check
We don't need strict typing as of now. This is only a factor when the
dependent_value is type boolean: false/true and the query params value is a
string.

We could change the type in the seed to string instead of boolean but honestly
I feel like this is fine. Maybe we can change it in the future if it's causing
issues for whatever reason.