Add rs_40_966 seed data
This was a huge one that brought up a lot of thought and some bugs.
The bugs were (hopefully) fully fixed in previous commits.

The thoughts include the direction of the project. I think I may want to change
the way the sentence is displayed. As the complexity of statutes increase I
think we need to display the full statute along with the sentence. Idk yet,
yolooooooo.